+++ src/main/java/um/tesoreria/core/hexagonal/persona/domain/model/DeudaExamen.java
@@ -1,0 +1,16 @@
+package um.tesoreria.core.hexagonal.persona.domain.model;
+
+import lombok.*;
+
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class DeudaExamen {
+
+    private Boolean autorizadoRendir;
+    private Boolean matriculaPagada;
+    private Integer cuotasAdeudadas;
+
+}