+++ src/main/java/um/tesoreria/core/hexagonal/setup/domain/model/Setup.java
@@ -1,0 +1,16 @@
+package um.tesoreria.core.hexagonal.setup.domain.model;
+
+import lombok.*;
+
+import java.math.BigDecimal;
+
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class Setup {
+    private Integer setupId;
+    private Integer cuotasPermitidas;
+    private BigDecimal cuentaHonorariosPagar;
+}