+++ src/main/java/um/tesoreria/core/hexagonal/setup/domain/ports/in/GetLastSetupUseCase.java
@@ -1,0 +1,8 @@
+package um.tesoreria.core.hexagonal.setup.domain.ports.in;
+
+import um.tesoreria.core.hexagonal.setup.domain.model.Setup;
+import java.util.Optional;
+
+public interface GetLastSetupUseCase {
+    Optional<Setup> getLastSetup();
+}