+++ src/main/java/um/tesoreria/core/hexagonal/setup/domain/ports/out/SetupRepository.java
@@ -1,0 +1,8 @@
+package um.tesoreria.core.hexagonal.setup.domain.ports.out;
+
+import um.tesoreria.core.hexagonal.setup.domain.model.Setup;
+import java.util.Optional;
+
+public interface SetupRepository {
+    Optional<Setup> findLast();
+}