+++ src/main/java/um/tesoreria/core/hexagonal/setup/infrastructure/web/dto/SetupResponse.java
@@ -1,0 +1,16 @@
+package um.tesoreria.core.hexagonal.setup.infrastructure.web.dto;
+
+import lombok.*;
+
+import java.math.BigDecimal;
+
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class SetupResponse {
+    private Integer setupId;
+    private Integer cuotasPermitidas;
+    private BigDecimal cuentaHonorariosPagar;
+}